requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
analytics-python==1.4.0
anyio==3.5.0
asgiref==3.5.0
async-timeout==4.0.2
attrs==21.4.0
backoff==1.10.0
bcrypt==3.2.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.0.4
colorama==0.4.4
cryptography==36.0.2
cycler==0.11.0
Cython==0.29.28
fastapi==0.75.0
ffmpy==0.3.0
fonttools==4.31.2
frozenlist==1.3.0
h11==0.13.0
idna==3.3
kiwisolver==1.4.1
linkify-it-py==1.0.3
markdown-it-py==2.0.1
MarkupSafe==2.1.1
matplotlib==3.5.1
mdit-py-plugins==0.3.0
mdurl==0.1.0
monotonic==1.6
multidict==6.0.2
numpy==1.22.3
opencv-python-headless
orjson==3.6.7
packaging==21.3
pandas==1.4.1
paramiko==2.10.3
Pillow==9.0.1
pycocotools==2.0.4
pycparser==2.21
pycryptodome==3.14.1
pydantic==1.9.0
pydub==0.25.1
PyNaCl==1.5.0
pyparsing==3.0.7
python-dateutil==2.8.2
python-multipart==0.0.5
pytz==2022.1
requests==2.27.1
six==1.16.0
sniffio==1.2.0
starlette==0.17.1
typing_extensions==4.1.1
uc-micro-py==1.0.1
urllib3==1.26.9
yarl==1.7.2
sklearn